Split huge snapp.go into smaller files
This branch creates snap_{local,remote,local_repo,remote_repo,yaml}
and utils.go from the huge catch-all snapp.go.

This is a first step to split them out into real pkgs.
